+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - add check the right path (hra-done)
 - add rack notif center (hra-done)
 - add order change customer and discount access (hra-done)
+- add new order non cash payment (hra-done)
 
 # 2.0.5
 ### features